refactor: move the cassandra id generation methods to CassandraLabelId
